+++ backend/alembic/versions/063a9e5bc287_create_user_table.py
@@ -29,7 +29,9 @@ def _create_users_table(db):
             CREATE TABLE users (
                 username VARCHAR(100) PRIMARY KEY,
                 password_hash VARCHAR(100),
-                salt VARCHAR(100)
+                salt VARCHAR(100),
+                logged_in BOOLEAN,
+                last_login TIMESTAMP
             );
         """)
