+++ backend/api/blueprints/__init__.py
@@ -1,0 +1,4 @@
+from .healthcheck import healthcheck_bp
+from .leads import leads_bp
+from .login import login_bp
+from .users import users_bp
