--- backend/main.py
@@ -1,4 +1,0 @@
-from api.app import app
-
-if __name__ == '__main__':
-    app.run(host='0.0.0.0')
