+++ backend/requirements-prod.txt
@@ -4,7 +4,6 @@ chardet==4.0.0
 click==7.1.2
 Flask==1.1.2
 Flask-Cors==3.0.10
-Flask-Login==0.5.0
 greenlet==1.1.0
 gunicorn==20.1.0
 idna==2.10
