+++ backend/run_dev.py
@@ -1,0 +1,4 @@
+from api.app import dev_app
+
+if __name__ == '__main__':
+    dev_app.run(host='0.0.0.0')
