+++ backend/tests/conftest.py
@@ -1,5 +1,2 @@
 import pytest
 
-
-@pytest.fixture
-def app
